fixed a couple minor bugs.  In the PME routines, we were accessing '0' indices in the tables if the scaled coordinates were exactly integer values. In the old code, we had a check within the 'create_scaled_direct_coordinates' subroutine to check that scaled coordinates weren't numerically zero, but now we update this check to make sure scaled coordinates aren't numerically integers (using 'mod' function)
we also fix the verlet list allocation for dilute systems, for which the density is so low that the verlet list isn't being allocated big enough.  Before, we had a floor
on the allocation of size total # of atoms, but this wasn't good enough because it didn't include a minimum number of neighbors.  So we define a minimum number of neighbors to be
'50', which is about 1/10 of what it would be in liquid water, and set the floor of the verlet list size to be total_atoms * min_neighbors