initial commit of NPT code
This commit has some bugs still to be worked out.

The value of min_neighbors was set to 200 in this commit, but this should
be changed later back to 50, and the verlet list should be reallocated on
volume changes.

A ctags file was added that allows code navigation in vim.